rename map component
