+++ spa/src/views/Map.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import Map from '../components/MapGL.vue';
+import Map from '../components/Map.vue';
 
 export default {
   name: 'MapView',
